basic parse & eval
parser uses core.match which I might be stretching but it seems to be fun so far

The problem with this core.match stuff is that I have to keep track of the rest of the tokens after I've successfully matched a phrase

so for example take the following source

```
let a = 1
let b = 2
```

when I match on the first statement its simple

```
(match tokens
  [{:type :let}, {:type :id, :value name}, {:type :=} & expr-tokens] (make-let name (parse-expr expr-tokens))
  ...
```

but I have to keep track of what tokens are left after the match.. this doesn't mutate the outer `tokens` variable.

So what I'm doing is each parser function returns a vec (tuple) of 2 items
1 - the ast node
2 - the rest of the tokens

It feels fine, but its kind of verbose to write out all the time

I also implemented a simple js emitter, this part is really easy because even in my ruby compiler it doesn't have mutable state.