add p- monad operators
parse array exprs

I've made a monad out of the parser result, but I'm not yet sold on its effectiveness

I have a few functions
- p-result, which is like pure in haskell
- p-into which lets you take the previous result & stick into a map at a specific key
- p-map, typical fmap
- p-then, typical bind/then
- p-pipe, takes a starting result & monadic operations.

All of this makes me feel really smart, but I've added ~30 lines of code & the code I've used it in some cases increased lines of code.

This kind of confirms some of my disposition against monads in general. The increase a lot of complexity & I'm not sure they solve that many problems, it only makes me feel clever.

The one nice thing is that p-result is a shared constructor so potentially its possible to change the data structure of the the parser result

But even then, I'm operating on the results raw in most the places, so unless I go all in, I don't get to reap the benefits.

I also don't plan on doing some lazy evaluation thing.

I'll keep it around for now & see if it grows into something useful. And if I have to get rid of it, its not a whole lot.